+++ src/scripts/create_global_mmlu.py
@@ -33,11 +33,18 @@ def main() -> None:
     # Define the dataset ID
     repo_id = "CohereLabs/Global-MMLU"
 
-    languages = ["uk", "el", "ro"]
+    languages = ["uk", "el", "ro", "sq"]
 
     for language in languages:
-        # Load the dataset
-        dataset = load_dataset(path=repo_id, name=language)
+        try:
+            # Load the dataset
+            dataset = load_dataset(path=repo_id, name=language)
+            lite = False
+        except ValueError:
+            # load the lite dataset
+            repo_id = "CohereLabs/Global-MMLU-lite"
+            dataset = load_dataset(path=repo_id, name=language)
+            lite = True
         assert isinstance(dataset, DatasetDict)
 
         # Convert the dataset to dataframes
@@ -50,46 +57,94 @@ def main() -> None:
         val_df = process_split(df=val_df, language=language)
         test_df = process_split(df=test_df, language=language)
 
-        # Create validation split from dev
-        val_size = 256
-        val_df = val_df.sample(n=val_size, random_state=4242, replace=False)
-        val_df = val_df.reset_index(drop=True)
-
-        # Create train and test splits from their test set
-        train_size = 1024
-        test_size = 2048
-
-        # Split their test into our train and test
-        train_df, final_test_df = train_test_split(
-            test_df, train_size=train_size, random_state=4242, stratify=test_df.category
-        )
-
-        # Sample our test set
-        final_test_df = final_test_df.sample(
-            n=test_size, random_state=4242, replace=False
-        )
-
-        # Reset the index
-        train_df = train_df.reset_index(drop=True)
-        final_test_df = final_test_df.reset_index(drop=True)
+        if lite:
+            train_df, val_df, test_df = make_mmlu_lite_splits(
+                val_df=val_df, test_df=test_df
+            )
+        else:
+            train_df, test_df, val_df = make_mmlu_splits(val_df=val_df, test_df=test_df)
 
         # Collect datasets in a dataset dictionary
         dataset = DatasetDict(
             {
                 "train": Dataset.from_pandas(train_df, split=Split.TRAIN),
                 "val": Dataset.from_pandas(val_df, split=Split.VALIDATION),
-                "test": Dataset.from_pandas(final_test_df, split=Split.TEST),
+                "test": Dataset.from_pandas(test_df, split=Split.TEST),
             }
         )
 
         # Create dataset ID
-        dataset_id = f"EuroEval/global-mmlu-{language}-mini"
-
+        lite_suffix = "-lite" if lite else ""
+        mini_suffix = "-mini" if not lite else ""
+        dataset_id = f"EuroEval/global-mmlu{lite_suffix}-{language}{mini_suffix}"
         # Remove the dataset from Hugging Face Hub if it already exists
         HfApi().delete_repo(dataset_id, repo_type="dataset", missing_ok=True)
 
         # Push the dataset to the Hugging Face Hub
         dataset.push_to_hub(dataset_id, private=True)
+
+
+def make_mmlu_lite_splits(
+    val_df: pd.DataFrame, test_df: pd.DataFrame
+) -> tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+    """Make splits for a lite dataset.
+
+    Args:
+        val_df: The validation dataframe.
+        test_df: The test dataframe.
+
+    Returns:
+        The final training, validation, and test dataframes.
+    """
+    val_size = 64
+    train_size = 128
+    train_df_final, test_df = train_test_split(
+        test_df, train_size=train_size, random_state=4242, stratify=test_df.category
+    )
+
+    val_df_final = val_df.sample(n=val_size, random_state=4242, replace=False)
+
+    # Use remaining val samples in the test split
+    remaining_val_df = val_df[~val_df.index.isin(val_df_final.index)]
+    test_df_final = pd.concat([test_df, remaining_val_df], ignore_index=True)
+
+    # Reset the index
+    train_df_final = train_df_final.reset_index(drop=True)
+    val_df_final = val_df_final.reset_index(drop=True)
+    test_df_final = test_df_final.reset_index(drop=True)
+
+    return train_df_final, val_df_final, test_df_final
+
+
+def make_mmlu_splits(
+    val_df: pd.DataFrame, test_df: pd.DataFrame
+) -> tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
+    """Make splits for a full dataset.
+
+    Args:
+        val_df: The validation dataframe.
+        test_df: The test dataframe.
+
+    Returns:
+        The final training, validation, and test dataframes.
+    """
+    val_size = 256
+    val_df_final = val_df.sample(n=val_size, random_state=4242, replace=False)
+
+    train_size = 1024
+    test_size = 2048
+
+    train_df_final, test_df = train_test_split(
+        test_df, train_size=train_size, random_state=4242, stratify=test_df.category
+    )
+
+    test_df_final = test_df.sample(n=test_size, random_state=4242, replace=False)
+
+    # Reset the index
+    train_df_final = train_df_final.reset_index(drop=True)
+    val_df_final = val_df_final.reset_index(drop=True)
+    test_df_final = test_df_final.reset_index(drop=True)
+    return train_df_final, test_df_final, val_df_final
 
 
 def process_split(df: pd.DataFrame, language: str) -> pd.DataFrame:
@@ -106,7 +161,7 @@ def process_split(df: pd.DataFrame, language: str) -> pd.DataFrame:
     df = filter_by_length(df=df)
     df = filter_repetitive(df=df)
     df = add_text_column(df=df, language=language)
-    df = df.loc[["text", "label", "category"]]
+    df = df.loc[:, ["text", "label", "category"]]
     df = df.drop_duplicates(inplace=False)
     df = df.reset_index(drop=True)
     return df