+++ src/scripts/create_multi_wiki_qa.py
@@ -55,6 +55,7 @@ LANGUAGES = [
     "hu",
     "ro",
     "ca",
+    "sq",
 ]