+++ src/scripts/create_wikiann.py
@@ -14,7 +14,7 @@ import pandas as pd
 from datasets import Dataset, DatasetDict, Split, load_dataset
 from huggingface_hub import HfApi
 
-LANGUAGE_CODES = ["lt", "lv", "hr", "bs", "ca"]
+LANGUAGE_CODES = ["lt", "lv", "hr", "bs", "ca", "sq"]
 
 
 def main() -> None: