+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added support for Albanian 🇦🇱! This includes the sentiment classification dataset
+  MMS-sq, the linguistic acceptability dataset ScaLA-sq, the named entity recognition
+  dataset WikiANN-sq, the reading comprehension dataset MultiWikiQA-sq, the
+  summarisation dataset LR-Sum-sq, the knowledge dataset Global-MMLU-Lite-sq,
+  and the common-sense reasoning dataset Winogrande-sq.
 - Added the Dutch common sense reasoning dataset COPA-NL, which is part of the
 Dutch [DUMB benchmark](https://github.com/wietsedv/dumb)